requirements.txt
fastapi
uvicorn
kafka-python
pydantic
python-dotenv
pytest
pytest-asyncio
aiohttp
httpx
requests